Use HMAC-256 rather than SHA1 for secret digests
HMAC is thought more resitant to hash extension or oracle attacks.  To
remedy this, use an HMAC with a constant key to fulfill the role of a
digest.

A constant (non-secret) key is used, "AttrVault", as a debugging aid so
one can quickly verify if some bytes are likely originated from
AttrVault.