Add AES-256-GCM Cryptor
This cipher suite is an authenticated encryption with associated data
(AEAD) one.  Keys that are prefixed with 'AES-256-GCM:' will use this
new Cryptor.

This is an abstraction and technique created after the Fernet coding,
which attr_vault borrows in part, was designed.  This predecessor
coding is AES-128-CBC with an extra HMAC, to attain authentication.

AES-256-GCM eliminates the HMAC as a separate cryptography primitive,
and the decryption and authentication check are thus simplified.  A
separate HMAC secret is no longer required, so the entire key text can
be used to expand the key width, to 256 bits, for AES-256.

An important attack vector to close is truncation of the
"authentication tag" (`auth_tag`), which can cause any message to
authenticate.  Its length is checked, and if it is not the
specification maximum (16), the ciphertext is rejected.  For more
details, see https://github.com/ruby/openssl/issues/63.